requirements.txt
asgiref==3.5.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==7.1.2
colorama==0.4.4
Django==4.0.3
django-environ==0.8.1
djangorestframework==3.13.1
google-api-core==2.7.1
google-api-python-client==1.12.11
google-auth==2.6.2
google-auth-httplib2==0.1.0
Google-Images-Search==1.4.2
googleapis-common-protos==1.56.0
httplib2==0.20.4
idna==3.3
Pillow==9.1.0
protobuf==3.20.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyfiglet==0.8.post1
pyparsing==3.0.7
python-resize-image==1.1.20
pytz==2022.1
requests==2.27.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
termcolor==1.1.0
tzdata==2022.1
uritemplate==3.0.1
urllib3==1.26.9
windows-curses==2.3.0
